LayoutTests/media/video-controls-transformed.html

<html>
<head>
  <title>Test controls on transformed video</title>
  <style type="text/css" media="screen">
    video {
      margin: 50px;
      -webkit-transform: rotate(20deg);
    }
  </style>
</head>
<body>
    <video controls></video>
    <p>Test controls on transformed video.</p>
    <p>This test only runs in DRT!</p>
    <script src=media-file.js></script>
    <script src=video-test.js></script>
    <script>
        testExpected("video.controls", null, '!=');
        waitForEvent('canplaythrough', function () {
            if (window.eventSender) {
                // Find the play button and click the middle of its bounding box.
                var playButton;
                var controlsShadow = layoutTestController.shadowRoot(video).firstChild.firstChild;
                for (child = controlsShadow.firstChild; child; child = child.nextSibling) {
                     if (layoutTestController.shadowPseudoId(child) == "-webkit-media-controls-play-button") {
                         playButton = child;
                         break;
                     }
                }

                if (!playButton) {
                    failTest("Failed to find play/pause button.");
                }
                var playButtonBoundingRect = playButton.getBoundingClientRect();
                var clickX = playButtonBoundingRect.left + playButtonBoundingRect.width / 2;
                var clickY = playButtonBoundingRect.top + playButtonBoundingRect.height / 2;
                eventSender.mouseMoveTo(clickX, clickY);
                eventSender.mouseDown();
                eventSender.mouseUp();
            }
            testExpected("video.paused", false);
            endTest();
        } );
        video.src = findMediaFile("video", "content/test");
    </script>
</body>
</html>
